Rename utils to kebab case
