+++ src/common/components/bet/BetSummary/betRow.svelte
