+++ src/common/components/bet/BetSummary/betSummaryTable.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   import type { BetSummary } from "$lib/interface/bet.types";
   import { betStore } from "$lib/stores/BetStore";
-  import { syncBetSummaryWithStore } from "$lib/utils/betCalculations";
-  import BetRow from "./BetRow.svelte";
-  import TableHeader from "./TableHeader.svelte";
+  import { syncBetSummaryWithStore } from "$lib/utils/bet-calculations";
+  import BetRow from "./betRow.svelte";
+  import TableHeader from "./tableHeader.svelte";
   import { createEventDispatcher } from "svelte";
   import { derived } from "svelte/store";
