+++ src/common/components/bet/BetSummary/tableHeader.svelte
