+++ src/common/components/bet/confirmPayment.svelte
