+++ src/common/components/bet/lotOptionalFilter.svelte
