+++ src/common/components/bet/lotteryTypeFilter.svelte
