+++ src/common/components/bet/lottoRulesModal.svelte
