+++ src/common/components/bet/numberPad.svelte
