+++ src/common/components/bet/selectedNumbers.svelte
