+++ src/common/components/bet/toggleMenu.svelte
