+++ src/common/components/navbar/coreNavbar.svelte
@@ -4,7 +4,7 @@
     import { onDestroy, onMount } from "svelte";
     import type { brandSetting } from "../../../interface/setting.type";
     import { goto } from "$app/navigation";
-    import { userAuth, userLogout } from "$lib/utils/userAuth";
+    import { userAuth, userLogout } from "$lib/utils/user-auth";
 
     let settings: brandSetting = {
         logo: { type: "text", value: "Loading..." },
