+++ src/components/lottery/lotteryResultCard.svelte
