+++ src/lib/utils/bet-utils.ts
