+++ src/lib/utils/user-auth.ts
